upgrade cleverhans/attacks/projected_gradient_descent.py

+++ cleverhans/attacks/projected_gradient_descent.py
@@ -92,7 +92,7 @@ class ProjectedGradientDescent(Attack):
     else:
       model_preds = self.model.get_probs(x)
       preds_max = tf.reduce_max(model_preds, 1, keepdims=True)
-      y = tf.to_float(tf.equal(model_preds, preds_max))
+      y = tf.cast(tf.equal(model_preds, preds_max), dtype=tf.float32)
       y = tf.stop_gradient(y)
       targeted = False
       del model_preds
